+++ src/components/Evolve/Evolve.tsx
@@ -1,14 +1,33 @@
 import React, { useCallback, useState } from 'react';
 import Modal from 'semantic-ui-react/dist/commonjs/modules/Modal';
 import Button from 'semantic-ui-react/dist/commonjs/elements/Button';
+import Radio from 'semantic-ui-react/dist/commonjs/addons/Radio';
 import useStore from 'store';
+import POKEMON from 'constants/pokemon';
+import { TEncounter } from 'constants/types';
 import styles from './Evolve.module.scss';
 
-const Evolve: React.FC = () => {
-  const [open, setOpen] = useState(false);
+interface EvolveProps {
+  encounter: TEncounter;
+  evolveIds: number[];
+}
+
+const Evolve: React.FC<EvolveProps> = ({ encounter, evolveIds }) => {
+  const changePokemon = useStore((state) => state.changePokemon);
   const darkMode = useStore(useCallback((state) => state.darkMode, []));
+  const [selected, setSelected] = useState(encounter.pokemon);
+  const [open, setOpen] = useState(false);
 
   const handleClose = () => {
+    setOpen(false);
+  };
+
+  const handleChange = (newId: number) => {
+    setSelected(newId);
+  };
+
+  const handleEvolve = () => {
+    changePokemon(encounter.id, selected);
     setOpen(false);
   };
 
@@ -29,10 +48,26 @@ const Evolve: React.FC = () => {
         />
       }
     >
-      <Modal.Header>Evolve Options</Modal.Header>
-      <Modal.Content>Evolve</Modal.Content>
+      <Modal.Header>Evolution Chain</Modal.Header>
+      <Modal.Content>
+        {evolveIds.map((id) => {
+          const foundPokemon = POKEMON.find((poke) => poke.value === id);
+          return (
+            <div key={`evolve-${id}`}>
+              <img src={foundPokemon?.image} alt={foundPokemon?.text} />{' '}
+              <Radio
+                checked={id === selected}
+                label={foundPokemon?.text}
+                onChange={(e, data) => handleChange(data.value as number)}
+                value={id}
+              />
+            </div>
+          );
+        })}
+      </Modal.Content>
       <Modal.Actions>
         <Button onClick={handleClose}>Cancel</Button>
+        <Button onClick={handleEvolve}>Save</Button>
       </Modal.Actions>
     </Modal>
   );